+++ models/trade.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
+from decimal import Decimal
 from typing import Optional
 
 class Trade(BaseModel):
@@ -8,17 +9,17 @@ class Trade(BaseModel):
     trade_date: str
     trade_time: str
     trade_timestamp: int
-    trade_price: float
-    trade_volume: float
+    trade_price: Decimal
+    trade_volume: Decimal
     ask_bid: str  # "ASK", "BID"
-    prev_closing_price: float
+    prev_closing_price: Decimal
     change: str  # "RISE", "FALL", "EVEN"
-    change_price: float
+    change_price: Decimal
     sequential_id: int
-    best_ask_price: float
-    best_ask_size: float
-    best_bid_price: float
-    best_bid_size: float
+    best_ask_price: Decimal
+    best_ask_size: Decimal
+    best_bid_price: Decimal
+    best_bid_size: Decimal
     stream_type: str
 
     class Config:
